Split 'platform' entrypoint into 'js' and 'web' entrypoints, move 'flag-parser' to 'sd-ce-pf' (and 'bundle') entrypoint only.
