+++ packages/webcomponentsjs/.gitignore
@@ -1,5 +1,6 @@
 custom-elements-es5-adapter.js
 webcomponents-bundle.js*
 bundles/**/*
+src/flag-parser.*
 src/platform/
 /buildcache
