+++ packages/webcomponentsjs/src/entrypoints/webcomponents-js-index.js
@@ -1,6 +1,6 @@
 /**
 @license
-Copyright (c) 2017 The Polymer Project Authors. All rights reserved.
+Copyright (c) 2020 The Polymer Project Authors. All rights reserved.
 This code may only be used under the BSD style license found at http://polymer.github.io/LICENSE.txt
 The complete set of authors may be found at http://polymer.github.io/AUTHORS.txt
 The complete set of contributors may be found at http://polymer.github.io/CONTRIBUTORS.txt
@@ -9,9 +9,5 @@ subject to an additional IP rights grant found at http://polymer.github.io/PATEN
 */
 
 import '../platform/es6-misc.js';
-import '../platform/custom-event.js';
 import '../platform/promise.js';
 import '../platform/symbol.js';
-import '../platform/flag-parser.js';
-import '../platform/baseuri.js';
-import '../platform/unresolved.js';
