+++ packages/webcomponentsjs/src/entrypoints/webcomponents-web-index.js
@@ -1,0 +1,13 @@
+/**
+@license
+Copyright (c) 2020 The Polymer Project Authors. All rights reserved.
+This code may only be used under the BSD style license found at http://polymer.github.io/LICENSE.txt
+The complete set of authors may be found at http://polymer.github.io/AUTHORS.txt
+The complete set of contributors may be found at http://polymer.github.io/CONTRIBUTORS.txt
+Code distributed by Google as part of the polymer project is also
+subject to an additional IP rights grant found at http://polymer.github.io/PATENTS.txt
+*/
+
+import '../platform/custom-event.js';
+import '../platform/baseuri.js';
+import '../platform/unresolved.js';
