+++ packages/webcomponentsjs/tsconfig.json
@@ -6,7 +6,7 @@
     "module": "esnext",
     "declaration": true,
     "sourceMap": true,
-    "outDir": "./src/platform",
+    "outDir": "./src",
     "importHelpers": true,
     "strict": true,
     "noImplicitAny": true,
@@ -23,6 +23,6 @@
     "moduleResolution": "node"
   },
   "include": [
-    "ts_src/platform"
+    "ts_src"
   ]
 }
